requirements.txt
reportportal-client==5.3.2
robotframework==6.0.2
robotframework-reportportal==5.3.2
numpy==1.24.3
pandas==2.0.1
PyYAML==6.0